P3/Xiutian/requirements.txt
fastapi
pydantic
sqlalchemy
alembic
uvicorn
psycopg2-binary
python-dotenv
httpx
pytest
